Change how SD is decided
Source Depth (SD) is now calculated by the code instead of manually entered. After the bathymetry for the first radial is generated, the depth at the site is extracted from that. 10m is subtracted from this to get SD (i.e. assuming the hydrophones are positioned 10m off the bottom).